feat: extract ConversationRepository from AppPersistence
Move conversation table schema, prepared statements, and CRUD methods
(getActiveConversation, createConversation, updateConversationSession,
updateConversationContext, archiveConversation, getRecentConversations,
incrementMessageCount) into a dedicated ConversationRepository module.

AppPersistence delegates to the repository via a factory function,
maintaining full backward compatibility for all consumers.
